[Add]無限に買えるように
