+++ .gitignore
@@ -4,3 +4,4 @@ npm-debug.log
 .env.sh
 .env
 .env-mac.sh
+index-mac.js
