--- app.yaml
@@ -1,2 +1,0 @@
-env: flex
-runtime: nodejs
